Refactored Menus.js to MenuManager.js
Also extracted DefaultMenus.js to do the brackets default Menu
initialisation
