+++ src/brackets.js
@@ -77,7 +77,8 @@ define(function (require, exports, module) {
         PerfUtils               = require("utils/PerfUtils"),
         FileIndexManager        = require("project/FileIndexManager"),
         QuickOpen               = require("search/QuickOpen"),
-        Menus                   = require("command/Menus"),
+        MenuManager             = require("command/MenuManager"),
+        DefaultMenus            = require("command/DefaultMenus"),
         FileUtils               = require("file/FileUtils"),
         MainViewHTML            = require("text!htmlContent/main-view.html"),
         Strings                 = require("strings"),
@@ -132,7 +133,7 @@ define(function (require, exports, module) {
             CommandManager          : CommandManager,
             FileSyncManager         : FileSyncManager,
             FileIndexManager        : FileIndexManager,
-            Menus                   : Menus,
+            MenuManager             : MenuManager,
             KeyBindingManager       : KeyBindingManager,
             CodeHintManager         : CodeHintManager,
             CSSUtils                : require("language/CSSUtils"),
